+++ spec/auction_fun_core/operations/user_context/authentication_operation_spec.rb
@@ -2,20 +2,16 @@
 
 require "spec_helper"
 
-RSpec.describe AuctionFunCore::Operations::UserContext::RegistrationOperation, type: :operation do
-  let(:user_repository) { AuctionFunCore::Repos::UserContext::UserRepository.new }
+RSpec.describe AuctionFunCore::Operations::UserContext::AuthenticationOperation, type: :operation do
+  let(:user_repo) { AuctionCore::Repos::UserRepo.new }
 
   describe ".call(attributes, &block)" do
     let(:operation) { described_class }
 
     context "when block is given" do
       context "when operation happens with success" do
-        let(:attributes) do
-          Factory.structs[:user]
-            .to_h
-            .except(:id, :created_at, :updated_at, :password_digest)
-            .merge!(password: "password", password_confirmation: "password")
-        end
+        let(:user) { Factory[:user] }
+        let(:attributes) { {login: user.email, password: "password"} }
 
         it "expect result success matching block" do
           matched_success = nil
@@ -44,51 +40,36 @@ RSpec.describe AuctionFunCore::Operations::UserContext::RegistrationOperation, t
           end
 
           expect(matched_success).to be_nil
-          expect(matched_failure[:name]).to include(I18n.t("contracts.errors.key?"))
+          expect(matched_failure[:login]).to include(I18n.t("contracts.errors.key?"))
         end
       end
     end
   end
 
-  describe "#call" do
+  describe "#call(attributes)" do
     subject(:operation) { described_class.new.call(attributes) }
 
-    context "when contract are not valid" do
+    context "when contract are invalid" do
       let(:attributes) { Dry::Core::Constants::EMPTY_HASH }
-
-      it "expect not persist new user on database" do
-        expect(user_repository.count).to be_zero
-
-        expect { operation }.not_to change(user_repository, :count)
-      end
 
       it "expect return failure with error messages" do
         expect(operation).to be_failure
-        expect(operation.failure[:name]).to include(I18n.t("contracts.errors.key?"))
+        expect(operation.failure[:login]).to include(I18n.t("contracts.errors.key?"))
       end
     end
 
     context "when contract are valid" do
-      let(:attributes) do
-        Factory.structs[:user]
-          .to_h
-          .except(:id, :created_at, :updated_at, :password_digest)
-          .merge!(password: "password", password_confirmation: "password")
-      end
+      let(:user) { Factory[:user] }
+      let(:attributes) { {login: user.email, password: "password"} }
 
       before do
         allow(AuctionFunCore::Application[:event]).to receive(:publish)
       end
 
-      it "expect persist new user on database and dispatch event registration" do
-        expect { operation }.to change(user_repository, :count).from(0).to(1)
+      it "expect return success" do
+        expect(operation).to be_success
 
         expect(AuctionFunCore::Application[:event]).to have_received(:publish).once
-      end
-
-      it "expect return success without error messages" do
-        expect(operation).to be_success
-        expect(operation.failure).to be_blank
       end
     end
   end