+++ spec/support/factories/users.rb
@@ -6,7 +6,7 @@ Factory.define(:user) do |f|
   f.phone { fake(:phone_number, :cell_phone_in_e164).tr_s("^0-9", "") }
   f.password_digest { BCrypt::Password.create("password") }
 
-  f.trait :disabled do |t|
+  f.trait :inactive do |t|
     t.active { false }
   end
 end